requirements.txt
fastapi>=0.115,<1.0
uvicorn>=0.30,<1.0
openai>=1.30,<2.0
python-multipart>=0.0.9,<1.0
pydantic>=2.7,<3.0
pypdf>=4.2,<6.0
python-docx>=1.1,<2.0
